Resolve warnings from circular dependencies
