+++ packages/react/src/elements/components/Typography/_constants.js
@@ -1,0 +1,24 @@
+export const _VALID_COLORS = Object.freeze([
+  "hig-white",
+  "hig-cool-gray-70",
+  "hig-blue-50",
+  "hig-green-good",
+  "hig-yellow-warning",
+  "hig-red-alert"
+]);
+
+export const _VALID_SIZES = Object.freeze(["small", "medium", "large"]);
+
+export const _VALID_TYPES = Object.freeze([
+  "h1",
+  "h2",
+  "h3",
+  "text",
+  // Types below intended to be deprecated
+  "sub1",
+  "sub2",
+  "body",
+  "bold",
+  "disabled",
+  "caption"
+]);
